Update permutator.php
I've tinkered with this enough to feel confident that increasing my recursion depth is enough to have essentially as much accuracy as I want. I'm still not sure about the pattern of recursion depth requirements, but the interval in depth requirement from the last array length to the next grows each step.