Update custom_vector.cpp
I've added a ! operator overload to the custom_vector_iterator class to check to see whether the class is currently referencing the last node of the vector that was passed to the constructor of the class when it was instantiated. This ! operator can be doubled to negate the result of the check and see whether the custom_vector_iterator object references a valid iteration point. This saves some typing compared to explicit comparison to the vector.end().