Create caesar_cipher2.c
This is a more advanced (and less efficient) way of implementing a caesar cipher. This uses a doubly linked/cyclic linked list with nodes associated with each unsigned character
value (0-255). Each character of a string to be encrypted or decrypted has its value located within the list. Next, the list is iterated forwards or backwards N times, as specified by the shift number. Since the list loops around, one can safely underflow from the value 0 to 253 when encrypting, for example, and overflow back to the original value when decrypting.

While this approach isn't practical (it is inferior to the earlier code for the same purpose), it serves as a good example of how/why one might use a linked list of this type.