Create read_line.c
This is yet another program that I wrote (paired with read_line.h for use as a library) to show someone how one might read variably sized input from stdin, safely. The reinvention of the common functions from string.h were just there as a conversational point, since a class this individual was taking disallowed the usage of the aforementioned header. This also allowed for me to introduce some new commandline syntax for compiling multiple source files, since that was something he'd not yet encountered in his classes.