Create json.cpp
This is a JSON interpreter and generator that I wrote for C++. I didn't like any of the available JSON solutions for C++. The most popular one that I noticed essentially analyzed existing JSON for its structure and generated a class tailored specifically for that JSON data hierarchy. Mine works like an unstructured map of any number of dimensions whose keys are type agnostic. The only caveat is that there's ambiguity between double and integer datatypes for the constant value, zero. This is remedied via explicit type-casting to invoke the correct subscript overload method.

Note that this code could be improved greatly as I was still familiarizing myself with C++ templating and operator overloading at the time of writing it. At some point, I may return to work on this. As it is, it passes the tests that I threw at it, but I haven't thrown any production data sets at it, as of yet.

This file pairs with json.h and json_test.cpp