Create json_test.cpp
This file was created to test the functionality of my JSON parsing/generating class as I built it. One should clearly see the drop in complexity of class usage and increase in versatility thereof as the tests increase in number. See json.cpp and json.h for more information regarding the purpose of this code.