Create primes.php
I was inspired by a Mathologer video on Youtube that showed Euler's method of calculating the number of unique partitions of an integer and an adaptation thereof, allowing for the sum of all factors of an integer, instead. I implemented this mathematical process in PHP because of the ease of string-based math using the bcmath module. Owing to the fact that one can get the sum of all factors of a given integer, I was able to check to see if the sum of an integer was equal to that intger plus 1. This means that the only factors of the integer in question are one and itself. This means it's prime. I used this method to detect all primes in the first N numbers (100,000 in the script as it is).

While I'm aware that this isn't a very efficient way of doing this, it was a fun challenge, and I've decided to upload it to share as a conversational piece for friends who are interested in this sort of thing.