Create single_list_reverse.cpp
This was in a coding challenge, and I thought I'd give it a try, when I saw it.
In reality, I'd always just make a doubly linked list, so that this sort of task would require no processing.
To be more precise, storing a doubly-linked list within a container (or a circular doubly-linked list) would remove the need for any of this list traversial and processing.

Example container with doubly linked list:
typedef struct NODE {
  unsigned char uc_id;

  NODE *lp_next;
  NODE *lp_prev;
} NODE;

typedef struct CONTAINER {
  NODE *lp_last_node;
  NODE *lp_first_node;
} CONTAINER;

With this, one could start on lp_last_node and iterate backwardly with lp_prev. Likewise, with a circular doubly-linked list, one could start on the first node's ->lp_prev parameter and iterate backwardly until the first node's ->lp_prev parameter is reached a second time.

I.E.
node = root ->lp_prev;
do {
  //Do something with the current node.
  node = node ->lp_prev;
} while ( node != root ->lp_prev );

This is a useless exercise, but it was still fun to figure out.