Create longest_palindromic_substring.cpp
This has two methods included within it. The first method I made was to split a string into all of its possible substring components and to check to see whether any of them are palindromic. For all palindromic substrings found, I checked to see which of them was the longest and returned that.

The next approach I took was to recursively check in two directions with a variable starting point for each of the directions and a boundary check for the leftward starting point.
The right-moving check in the recursive call is already bounded by a null terminator at the end of the string. To make this work with binary data, an end point would need to be passed rather than checking for inequality between a and b to avoid a buffer overrun.