Create map.c
This is a custom map implementation that can have quick lookup while also maintaining the original insertion order by keeping an array of references to new nodes. I need to allow for prepending the list as well as the current appending method so that index based subscript access within the map is more feature rich. I'm also thinking about optimizing the searches within the tranches by hashing and/or alphabetically sorting the keys stored within them.