Create suffix_tree.cpp
This was an interesting challenge to develop based upon a visual strategy in a Youtube video. Once I finish making all of the optimizations that I have in mind, I'll take a look at others' code to see how similar it is and whether or not I've missed any possible optimizations.