Create bit_field.c
I was testing bit fields within structures in C with GCC because I had only ever done bit manipulations within single variables or arrays thereof, in the past. It doesn't 
seem that there is any great advantage for bit fields unless one-byte alignment and manual dummy variable padding is used to pack the structures. In this case, it can improve code readability when compared to a series of bit shifts and binary OR operations to pack the necessary bits.