Update bit_field.c
I've updated the comments so as better to reflect the results of my testing, now that I've identified differences in implementation of bit fields within structures between MSVC and GCC. Bit fields seem always to be padded to at least one byte in length rather than having consecutive bit fields not ending on a multiple of 8 being crammed together on MSVC, as one would hope for/expect. On GCC, with packing, this works as I would expect. I'm definitely going to stick to manual bit field manipulation for cross-platform compliance, henceforth. What a waste of time and poorly standardized/implemented language feature.